+++ dags/geo_title_counts.py
@@ -2,6 +2,7 @@ import csv
 import logging
 import os
 from datetime import datetime
+import pandas as pd
 
 from airflow import DAG
 from airflow.hooks import S3Hook
@@ -12,6 +13,7 @@ from utils.s3 import upload
 from utils.nlp import NLPTransforms
 from datasets import job_postings
 from algorithms.aggregators.title import GeoTitleAggregator
+from algorithms.jobtitle_cleaner.clean import JobTitleStringClean, aggregate
 from config import config
 
 default_args = {
@@ -72,4 +74,69 @@ class GeoTitleCountOperator(BaseOperator):
         upload(s3_conn, count_filename, config['output_tables']['s3_path'])
         upload(s3_conn, rollup_filename, config['output_tables']['s3_path'])
 
-GeoTitleCountOperator(task_id='geo_title_count', dag=dag)
+class JobTitleCleanOperator(BaseOperator):
+    def execute(self, context):
+        s3_conn = S3Hook().get_conn()
+        quarter = datetime_to_quarter(context['execution_date'])
+
+        cleaned_count_filename = '{}/cleaned_geo_title_count_{}_new.csv'.format(
+            output_folder,
+            quarter
+        )
+
+        cleaned_rollup_filename = '{}/cleaned_title_count_{}.csv'.format(
+            output_folder,
+            quarter
+        )
+
+        count_filename = '{}/geo_title_count_{}.csv'.format(
+            output_folder,
+            quarter
+        )
+
+        rollup_filename = '{}/title_count_{}.csv'.format(
+            output_folder,
+            quarter
+        )
+
+        logging.info('Cleaning and aggregating geo job titles on %s', quarter)
+        geo_title_count_df = pd.read_csv(count_filename, header=None)
+        geo_title_count_df.columns = ['geo', 'title', 'count']
+        cleaned_geo_title_count_df = JobTitleStringClean().clean(geo_title_count_df)
+        agg_cleaned_geo_title_count_df = aggregate(cleaned_geo_title_count_df, ['geo', 'title'])
+
+        logging.info('Cleaning and aggregating job titles on %s', quarter)
+        title_count_df = pd.read_csv(rollup_filename)
+        title_count_df.columns = ['title', 'count']
+        cleaned_title_count_df = JobTitleStringClean().clean(title_count_df)
+        agg_cleaned_title_count_df = aggregate(cleaned_title_count_df, ['title'])
+
+        total_counts = 0
+        with open(cleaned_count_filename, 'w') as count_file:
+            clean_geo_writer = csv.writer(count_file, delimiter=',')
+            for idx, row in agg_cleaned_geo_title_count_df.iterrows():
+                total_counts += row['count']
+                clean_geo_writer.writerow([row['geo'], row['title'], row['count']])
+
+        rollup_counts = 0
+        with open(cleaned_rollup_filename, 'w') as count_file:
+            clean_writer = csv.writer(count_file, delimiter=',')
+            for idx, row in agg_cleaned_title_count_df.iterrows():
+                rollup_counts += row['count']
+                clean_writer.writerow([row['title'], row['count']])
+
+        logging.info(
+            'Found %s count rows and %s title rollup rows for %s',
+            total_counts,
+            rollup_counts,
+            quarter,
+        )
+
+        upload(s3_conn, cleaned_count_filename, config['output_tables']['s3_path'])
+        upload(s3_conn, cleaned_rollup_filename, config['output_tables']['s3_path'])
+
+jobtitle_clean = JobTitleCleanOperator(task_id='clean_title_count', dag=dag)
+geo_count = GeoTitleCountOperator(task_id='geo_title_count', dag=dag)
+
+jobtitle_clean.set_upstream(geo_count)
+
