+++ missed.txt
@@ -107040,3 +107040,263 @@ MD/lutherville timonium
 IN/wheatfield
 ME/winterport
 MO/cairo
+NY/long island
+NJ/morristown
+CA/north sacramento
+OH/north jackson
+TX/new caney
+MI/redford
+MI/goodison
+MI/redford
+NJ/monroe township
+CA/buttonwillow
+TX/vealmoor
+CA/van nuys
+NY/bayside
+VA/prince george
+NJ/sicklerville
+NY/forest hills
+NY/melrose
+CA/sherman oaks
+NY/forest hills
+NY/morris heights
+NY/edgewood
+TX/dallas/fort worth
+NY/long island
+VA/fort eustis
+FL/tampa-st petersburg
+OK/hinton
+TX/shamrock
+OK/waynoka
+KY/calvert city
+NY/long island
+KS/countryside
+TX/bluff springs
+KS/countryside
+OH/north jackson
+TX/dallas/fort worth
+OH/sylvania township
+CA/panorama city
+IN/kentland
+IL/lincoln park
+IL/lincoln park
+CA/chatsworth
+CA/chatsworth
+CT/unionville
+CA/ventura
+CA/chatsworth
+CA/north hollywood
+CA/ventura
+CA/north hollywood
+  /kiev
+NJ/west berlin
+PA/moon township
+OH/amlin
+MD/kent island
+TX/dallas/fort worth
+NC/raleigh-durham
+NY/long island city
+NC/raleigh-durham
+NJ/pedricktown
+IN/castleton
+MD/annapolis junction
+NY/westchester
+NY/westchester
+NC/raleigh-durham
+AR/hot springs national park
+MD/woolford
+NY/gpo
+WA/lynwood
+MI/canton
+NC/boogertown
+PA/newtown square
+CA/ventura
+CA/orange county
+FL/sand lake
+NC/raleigh durham
+CA/ventura
+NC/raleigh-durham
+CA/ventura
+CA/orange county
+PA/cranberry township
+PA/glen mills
+PA/wexford
+PA/wexford
+PA/wexford
+AZ/rincon
+FL/land o lakes
+MA/watertown
+NJ/sewell
+KY/simpsonville
+VA/east hampton
+MN/west coon rapids
+WV/weirton heights
+TN/cordova
+ME/west scarborough
+ME/south portland gardens
+NY/east new york
+NH/east concord
+NY/west albany
+MO/lees summit
+NV/carson city
+MT/butte
+MS/west gulfport
+MT/anaconda
+MA/south boston
+MO/east independence
+NH/derry village
+FL/hastings
+FL/hastings
+MD/woolford
+PA/south mountain
+PA/jamison
+IN/roselawn
+IA/woodward
+OH/flat rock
+FL/enterprise
+NJ/hewitt
+NY/forest hills
+MD/hunt valley
+MD/marlboro
+CA/city of industry
+SC/hilton head
+CA/woodland hills
+MI/bloomfield
+AL/tanner
+VA/washington dc
+MA/north grafton
+VA/washington dc
+MA/north quincy
+NY/midwood
+CA/city of industry
+PA/butztown
+IA/pocahontas
+VT/fairfax
+MN/royalton
+CA/canyon country
+CA/sherman oaks
+NY/latham
+MD/hunt valley
+CA/sherman oaks
+MD/hunt valley
+CA/sherman oaks
+CA/sherman oaks
+GA/brookhaven
+CA/sherman oaks
+CT/yalesville
+FL/tampa bay
+VT/north burlington
+CA/cowan heights
+WI/randolph
+WA/times square
+IN/crane
+MD/hanover
+IN/crane
+IN/crane
+IN/crane
+IA/iowa state university
+VA/tysons
+TN/antioch
+TN/madison
+NJ/basking ridge
+FL/metro market
+NY/long island
+AK/juneau
+VA/dulles
+MT/butte
+PA/wilkes barre
+NW/nationwide
+FL/tampa-st petersburg
+AK/juneau
+MT/butte
+PA/wilkes barre
+MT/butte
+NC/winston salem
+NC/winston salem
+NC/winston salem
+NC/winston salem
+NC/winston salem
+NY/long island
+AK/juneau
+NC/winston salem
+PA/wilkes barre
+AK/juneau
+PA/wilkes barre
+ME/gouldsboro
+MN/tyler
+WY/midwest
+WY/midwest
+MT/butte
+PA/wilkes barre
+AK/juneau
+NJ/morristown
+NJ/basking ridge
+NJ/basking ridge
+NC/arden
+CA/woodland hills
+NC/arden
+NY/midtown
+NY/long island
+NY/long island
+NY/long island
+NY/long island
+NY/long island
+NY/long island
+MA/north easton
+FL/tampa-st petersburg
+VT/randolph
+PA/wilkes barre
+FL/tampa-st petersburg
+MN/nisswa
+CA/ventura
+FL/tampa bay
+WI/markesan
+NJ/bergen
+NJ/morristown
+NY/westchester
+CA/van nuys
+VA/hampton roads
+CA/van nuys
+NY/westchester county
+CA/chatsworth
+CA/chatsworth
+CA/chatsworth
+NJ/cherry hill township
+NJ/lawrence township
+NJ/lawrence township
+NY/long island
+NY/westchester
+NY/westchester
+NY/westchester
+DE/christiana
+GA/warm springs
+NV/carson city
+TX/dallas/fort worth
+TX/dallas/fort worth
+IL/haymarket
+LA/cow island
+TX/dallas/fort worth
+MI/windsor
+NY/long island
+DE/christiana
+DE/christiana
+TX/cypress
+MA/wellesley hills
+CA/la jolla
+CA/orange county
+NY/westchester
+NC/raleigh-durham
+MA/wellesley hills
+MA/franklin
+NC/raleigh-durham
+MA/needham heights
+TN/old hickory
+MO/west county
+FL/tampa bay
+CA/northridge
+SC/hilton head
+NJ/central
+NY/glenmont
+MD/lutherville timonium
+NJ/morristown
+NY/westchester
